refactor: 위도 경도 Long 타입에서 BigDecimal로 변경

+++ src/main/java/com/econo_4factorial/newproject/base/service/weather/WeatherFeignClient.java
@@ -5,13 +5,15 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "weatherClient", url = "https://api.openweathermap.org/data/2.5")
 public interface WeatherFeignClient {
 
     @GetMapping("/weather")
     WeatherRes getWeather(
-            @RequestParam("lat") Double latitude,
-            @RequestParam("lon") Double longitude,
+            @RequestParam("lat") BigDecimal latitude,
+            @RequestParam("lon") BigDecimal longitude,
             @RequestParam("appid") String apikey
     );
 }
